Just preserving a version that seems to work on eddy with pgf90 12.4-0 64-bit target on x86-64 Linux -tp sandybridge H5_PACKAGE_VERSION "1.8.15-patch1"
This same version CRASHES with gfortran on my Mac, and on cheyenne with

Currently Loaded Modules:
  1) ncarenv/1.1           4) mpt/2.15f        7) matlab/R2016b  10) python/2.7.13
  2) intel/16.0.3          5) netcdf/4.4.1.1   8) nco/4.6.9      11) xxdiff/4.0.1
  3) ncarcompilers/0.4.1   6) git/2.10.2       9) ncview/2.1.7   12) ncl/6.4.0



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps@11961 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a4ff9bf40520f0c8d0e0f3d3f18a8686a59bac06

+++ assimilation_code/modules/assimilation/assim_tools_mod.f90
@@ -139,7 +139,7 @@ real(r8) :: cutoff                          = 0.2_r8
 logical  :: sort_obs_inc                    = .false.
 logical  :: spread_restoration              = .false.
 logical  :: sampling_error_correction       = .false.
-integer  :: adaptive_localization_threshold = -1
+integer(i8) :: adaptive_localization_threshold = -1
 real(r8) :: adaptive_cutoff_floor           = 0.0_r8
 integer  :: print_every_nth_obs             = 0
